Split algorithm and functional utils into their own files. Add is_standard_layout and is_trivial type traits
