+++ library/include/rocwmma/internal/layout/matrix_layout_impl.hpp
@@ -28,6 +28,7 @@
 
 #include "layout.hpp"
 #include "layout_traits.hpp"
+#include "../utility/algorithm.hpp"
 
 namespace rocwmma
 {
