+++ library/include/rocwmma/internal/utility/algorithm_impl.hpp
@@ -1,0 +1,51 @@
+/*******************************************************************************
+ *
+ * MIT License
+ *
+ * Copyright (C) 2021-2025 Advanced Micro Devices, Inc. All rights reserved.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ *
+ *******************************************************************************/
+
+#ifndef ROCWMMA_UTILITY_ALGORITHM_IMPL_HPP
+#define ROCWMMA_UTILITY_ALGORITHM_IMPL_HPP
+
+namespace rocwmma
+{
+    namespace detail
+    {
+        // TODO: Separate file?
+        template <typename T>
+        ROCWMMA_HOST_DEVICE constexpr decltype(auto) max(T&& a, T&& b)
+        {
+            return (a < b) ? b : a;
+        }
+
+        template <typename T>
+        ROCWMMA_HOST_DEVICE constexpr decltype(auto) min(T&& a, T&& b)
+        {
+            return (a < b) ? a : b;
+        }
+
+    } // namespace detail
+
+} // namespace rocwmma
+
+#endif // ROCWMMA_UTILITY_ALGORITHM_IMPL_HPP
