+++ library/include/rocwmma/internal/utility/type_traits.hpp
@@ -66,11 +66,8 @@ namespace rocwmma
     using detail::is_same_v;
     using detail::is_signed;
     using detail::is_signed_v;
-
-    // TODO: override namespace not detail
-    using __hip_internal::is_standard_layout;
-    using __hip_internal::is_trivial;
-
+    using detail::is_standard_layout;
+    using detail::is_trivial;
     using detail::is_void;
     using detail::is_void_v;
     using detail::remove_const;
@@ -85,18 +82,6 @@ namespace rocwmma
     using detail::remove_volatile_t;
     using detail::true_type;
 
-    // TODO: goes into algorithm
-    using detail::max;
-    using detail::min;
-
-    // TODO: goes into functional
-    using detail::logical_or;
-    //using detail::logical_or_v;
-    using detail::logical_and;
-    //using detail::logical_and_v;
-    using detail::logical_not;
-    //using detail::logical_not_v;
-
 } // namespace rocwmma
 
 #define ROCWMMA_TYPE_TRAITS_IMPL_NAMESPACE rocwmma::detail
@@ -104,10 +89,6 @@ namespace rocwmma
 #else
 
 #include <type_traits>
-
-// TODO: move to own files
-#include <algorithm>
-#include <functional>
 namespace rocwmma
 {
     // std implementations
@@ -159,18 +140,6 @@ namespace rocwmma
     using std::remove_volatile;
     using std::remove_volatile_t;
     using std::true_type;
-
-    // TODO: goes into algorithm
-    using std::max;
-    using std::min;
-
-    // TODO: goes into functional
-    using std::logical_or;
-    //using std::logical_or_v;
-    using std::logical_and;
-    //using std::logical_and_v;
-    using std::logical_not;
-    //using std::logical_not_v;
 
 } // namespace rocwmma
