+++ library/include/rocwmma/internal/vector_util_impl.hpp
@@ -29,6 +29,7 @@
 
 #include "blend.hpp"
 #include "types.hpp"
+#include "utility/algorithm.hpp"
 #include "utility/vector.hpp"
 
 namespace rocwmma
